refactor renamed app.crud.dog -> dog_crud, app.crud.user -> user_crud
