+++ app/crud/__init__.py
@@ -1,5 +1,5 @@
-from .dog import dog
-from .user import user
+from .dog_crud import dog
+from .user_crud import user
 
 
 __all__ = ['dog', 'user']
